+++ crates/vm/levm/src/opcode_handlers/exchange.rs
@@ -9,14 +9,14 @@ use crate::{
 
 impl<'a> VM<'a> {
     // SWAP operation
-    pub fn op_swap(&mut self, depth: usize) -> Result<OpcodeResult, VMError> {
+    pub fn op_swap<const N: usize>(&mut self) -> Result<OpcodeResult, VMError> {
         let current_call_frame = self.current_call_frame_mut()?;
         current_call_frame.increase_consumed_gas(gas_cost::SWAPN)?;
 
-        if current_call_frame.stack.len() < depth {
+        if current_call_frame.stack.len() < N {
             return Err(ExceptionalHalt::StackUnderflow.into());
         }
-        current_call_frame.stack.swap(depth)?;
+        current_call_frame.stack.swap(N)?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }